refactor: refatora paths baseado em sistema operacional, organiza e otimiza arquivos e pastas
