+++ src/commands/compress-video/index.ts
@@ -1,34 +1,11 @@
 import chalk from 'chalk';
 import inquirer from 'inquirer';
-import loading from "../../utils/loading";
 import fs from 'fs';
 import { removeSpecialCharacters } from './utils/removeSpecialCharacters';
 import { supportedFiles } from './utils/supportedFiles';
-import { platform } from 'os';
-import { OperatingSystemPlatforms } from './utils/operatingSystem';
+import { platformOS, slashFormatByPlatformOS } from './utils/operatingSystem';
+import { compressVideo } from './utils/compressVideo';
 
-const ffmpegPath = require('@ffmpeg-installer/ffmpeg').path;
-const ffmpeg = require('fluent-ffmpeg');
-ffmpeg.setFfmpegPath(ffmpegPath);
-
-const platformOS = platform() as OperatingSystemPlatforms;
-
-async function compressVideo(inputPath: string, outputPath: string) {
-  const compressingLoader = loading("");
-
-  ffmpeg(inputPath)
-    .outputOptions([])
-    .save(outputPath)
-    .on('start', () => compressingLoader.start('Processando e comprimindo o vídeo...'))
-    .on('end', () => {
-      compressingLoader.succeed(chalk.greenBright('✅ Vídeo comprimido com sucesso!'));
-      console.log(chalk.green(`\n O vídeo comprimido foi salvo em: ${outputPath}`));
-    })
-    .on('error', (error: any) => {
-      compressingLoader.fail(chalk.red(error.message));
-    })
-    
-};
 
 export default async function compressVideoPrompts() {
   console.log(chalk.whiteBright('👇 Siga os passos abaixo para comprimir o vídeo! \n'));
@@ -42,15 +19,15 @@ export default async function compressVideoPrompts() {
         const inputPath = removeSpecialCharacters(input, platformOS);
         const fileExtension = inputPath.split('.').pop();
 
-        if (!(fileExtension && fileExtension in supportedFiles)) return `${chalk.red('O arquivo escolhido através deste diretório não é um tipo de arquivo suportado!')}. Insira um arquivo MOV ou MP4! (.mov ou .mp4 ao final do diretório)`;        
+        if (!(fileExtension && fileExtension in supportedFiles)) return `${chalk.red('O arquivo escolhido através deste diretório não é um tipo de arquivo suportado!')}. Insira um arquivo MOV ou MP4! (.mov | .mp4 no fim do nome do arquivo)}`;        
         if (!fs.existsSync(inputPath)) return `${chalk.red('Não foi possível encontrar o diretório informado do vídeo')}. Digite o diretório corretamente!`;        
         return true;
       },
     });
 
     const cleanInputPath = removeSpecialCharacters(inputPath, platformOS);
-    const filePath = cleanInputPath.split("/").slice(0, -1).join("/");
-    const oldFileName = cleanInputPath.split("/").pop()?.split(".").shift();
+    const filePath = cleanInputPath.split(slashFormatByPlatformOS[platformOS]).slice(0, -1).join(slashFormatByPlatformOS[platformOS]);
+    const oldFileName = cleanInputPath.split(slashFormatByPlatformOS[platformOS]).pop()?.split(".").shift();
     
     const { outputPath } = await inquirer.prompt({
       type: 'input',
@@ -83,7 +60,7 @@ export default async function compressVideoPrompts() {
     //   ],
     // });
     
-    const outputFile = `${outputPath}/${fileName}.${fileExtension}`.split('/').join('/');
+    const outputFile = `${outputPath}${slashFormatByPlatformOS[platformOS]}${fileName}.${fileExtension}`
     await compressVideo(cleanInputPath, outputFile);
   } catch (err: any) {
     console.error(chalk.red('❌ Ocorreu um erro:'), err.message);
