+++ src/commands/compress-video/utils/operatingSystem.ts
@@ -1,7 +1,17 @@
+import { platform } from 'os';
+
+export type OperatingSystemPlatforms = "win32" | "darwin" | "linux";
+
 export enum OperatingSystem {
   MAC = "darwin",
   LINUX = "linux",
   WIN = "win32"
 }
 
-export type OperatingSystemPlatforms = "win32" | "darwin" | "linux";
+export const platformOS = platform() as OperatingSystemPlatforms;
+
+export const slashFormatByPlatformOS = {
+  [OperatingSystem.LINUX]: '/',
+  [OperatingSystem.MAC]: '/',
+  [OperatingSystem.WIN]: '\\'
+}
