+++ src/commands/compress-video/utils/supportedFiles.ts
@@ -1,6 +1,7 @@
 export const supportedFiles = {
   mp4: "mp4",
   mov: "mov",
+  mkv: "mkv"
   // png: "png",
   // jpeg: "jpeg",
   // jpg: "jpg",
